+++ mm2src/coins/tendermint/htlc/nucleus/htlc.rs
@@ -12,15 +12,6 @@ use std::convert::TryFrom;
 pub(crate) const NUCLEUS_CREATE_HTLC_TYPE_URL: &str = "/nucleus.htlc.MsgCreateHTLC";
 pub(crate) const NUCLEUS_CLAIM_HTLC_TYPE_URL: &str = "/nucleus.htlc.MsgClaimHTLC";
 
-#[allow(dead_code)]
-pub(crate) struct NucleusHtlc {
-    /// Generated HTLC's ID.
-    pub(crate) id: String,
-
-    /// Message payload to be sent
-    pub(crate) msg_payload: cosmrs::Any,
-}
-
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub(crate) struct MsgCreateHtlc {
     /// Sender's address.
